[libc] major refactor of startup library
* separate initialization routines into _start and do_start for all
  architectures.
* lift do_start as a separate object library to avoid code duplication.
* (addtionally) address the problem of building hermetic libc with -fstack-pointer-*
